Implement auto scoped logging
Now there is a `logger` import that will handle setting the scope for
log the plugin (or root).  Converted all the `print` calls to use this
logger.  This can probably be cleaned up, but currently the scope is
tracked along with the listeners so that the context manager can be
called to convert the calls to use the scope.

Addresses #1

NOTE: still need to figure out configuration for the logging system,
right now it just sets it up in a static function `setup`