Reorganize a badly placed helper function
The `merge` method on the test helper was misplaced and it has been
moved into the utils functions.  This is then used to allow for
configuring which storage backend to use in db config, defaulting to
`anydbm` but allowing for `shelve` to be used.  This will be expanded to
include some of the others at some point.