Initial pass at plugin encapsulation
All plugins are imported within a `Plugin` wrapper object which contains
references to all the hooks of the plugin.  All hooks now are wrapped in
a `Hook` or subclass of that.  This gives control over the general
reference and can allow for on the fly disabling and loading.

Things to address:
* move the logging work into the Plugin system
* give attribute access to the plugin module from the wrapping class
* overload the module reference for the plugin with the wrapper