Add ability for PM based filters
You can now specify a channel type as the channel target of a filter and
it will apply the filter to any message on a channel of that type (this
is mostly for IM messages).  Had to rework the filter matching and now
use a dictionary for the different groupings.

This closes #20

Also added some guards for double closing databases and for capture
filters to use pipes for the regex.