About plugin build out
The about plugin now provides better information on the loaded plugins
using the docstrings for filters and the plugin modules.  There is some
wrapping done with docstrings for `Plugin` and `Hook` objects with
prefixes for context.  Docstrings were added to the plugins to give
context on what they are.

This addresses #18.